fix: change time_spl from unsigned int to double for WMX3 v3.7 SDK compatibility
WMX3 SDK v3.7 changed the StartCSplinePos API signature from
unsigned int* to double* for the time array parameter.

+++ wmx_ros2_package/src/follow_joint_trajectory_server.cpp
@@ -44,7 +44,7 @@ private:
   AdvancedMotion wmx3LibAm_;
   AdvMotion::PointTimeSplineCommand spl;
   AdvMotion::SplinePoint pt_spl[MAX_TRAJ_POINTS];
-  unsigned int time_spl[MAX_TRAJ_POINTS];
+  double time_spl[MAX_TRAJ_POINTS];
   AxisSelection axisSel;
   Io Wmx3Lib_Io_;
